requirements.txt
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
cachetools==4.2.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snscrape==0.4.3.20220106
sortedcontainers==2.4.0
soupsieve==2.7
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
python-telegram-bot==22.3
flask
requests
requests_oauthlib
python-dotenv
